Create Video Element.html
To add a video to the HTML Page, follow the steps below:

Create a video element in the HTML file with controls attribute.
Add an id with text video to the video element
Add a source element inside the video element with the src attribute pointing to a sample video and the type attribute set to video/mp4.
To make this video responsive, we will the width property of the video element to 100% and height property to auto. These would auto adjust the width and the height of the video based on the available width in the current viewport of the page.