Create Inline vs. Block Elements.html
Create a block element with the ID "block-example" - Create a <div> element with the ID block-example. <div> is a commonly used block element in HTML.

Create an inline element with the ID "inline-example" - Create a <span> element with the ID inline-example. <span> is a commonly used inline element in HTML.

Change the display property of the block element to inline - Make sure to write inline CSS rules in the <style> tag inside the <head> element. Change the display property of the block-example element to inline.

Change the display property of the inline element to block - Make sure to write inline CSS rules in the <style> tag inside the <head> element. Change the display property of the inline-example element to block.

Create an inline-block element with the ID "inline-block-example" - Create a <div> element with the ID inline-block-example, and using CSS in the <style> tag, change its display property to inline-block