Create Position Property.css
Create a div container with id container which has a width of 300px and height of 300px, and a border to see its boundary

Create a class box which holds the sets the width and height to 50px. Add the box class to all the three boxes that you create inside #container element

Inside the #container div, create a div with id 'static-box' and add background color blue. This box will not use any position property

Create another div with id relative-box inside the #container div having properties, background color to green, and position it relatively 20px from the top and 20px from the left of the #container div

Create a third div with id absolute-box,set the background color to red, and position it absolutely at 220px from the top and 20px from the left of the container div