Create JavaScript function.js
Create an HTML page with a div element having an id 'container'. Inside the 'container', create 6 div elements with class 'box'.

Apply a CSS rule to the '.box' class elements to make them display in columns with a column-count of 2.

Select the 3rd div element with class 'box' and make it span across 2 columns.

Write a JavaScript function changeColumnCount that takes an integer as an argument and sets the column-count of the div with id 'container' to that value.